refactor: BookmarkContent 내부 컴포넌트들 공통에서 사용할 수 있도록 추출
